requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
typing_extensions==4.12.2
yarl==1.17.1
yt-dlp==2024.11.4
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
typing_extensions==4.12.2
yarl==1.17.1
yt-dlp==2024.11.4
